extract dlog and deprecate self created editor
